# spec/openapi.yaml
openapi: 3.0.1
info:
  title: ZapMe
  description: Open source application to control things
  termsOfService: https://zapme.app/tos
  contact:
    name: Contact me
    url: https://zapme.app/contact
  license:
    name: 'The Apache License, Version 2.0'
    url: https://zapme.app/license
  version: v1
servers:
  - url: https://api.zapme.app
paths:
  /api/v1/account:
    post:
      tags:
        - Account
      summary: Create a new account
      operationId: CreateAccount
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CreateAccount'
              description: Message sent to server to create a new account
      responses:
        '201':
          description: Created account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOk'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
    delete:
      tags:
        - Account
      summary: Delete currently logged in account
      operationId: DeleteAccount
      parameters:
        - name: password
          in: header
          schema:
            maxLength: 256
            minLength: 10
            type: string
            example: Hq2yP1B^Fho&zRHxHkEu
        - name: reason
          in: header
          schema:
            maxLength: 1024
            minLength: 0
            type: string
      responses:
        '200':
          description: Empty
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
    get:
      tags:
        - Account
      summary: ''
      operationId: GetAccount
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
  /api/v1/account/recover-confirm:
    post:
      tags:
        - Account
      summary: Complete the password reset request using the token that was received in the users email
      operationId: AccountRecoveryConfirm
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RecoveryConfirm'
              description: Request sent to server to commit a password reset
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
  /api/v1/account/email:
    put:
      tags:
        - Account
      summary: Updates the account email
      operationId: UpdateEmail
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UpdateEmail'
              description: Request sent to server to update account email address
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateEmailOk'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
  /api/v1/account/pfp:
    put:
      tags:
        - Account
      summary: Updates the account profile picture
      operationId: UpdateProfilePicture
      parameters:
        - name: Hash-Sha256
          in: header
          description: '[Optional] Sha-256 hash of the image bytes to verify the integrity of the image server-side'
          schema:
            type: string
      requestBody:
        content:
          image/png:
            schema:
              type: string
              description: payload
              format: binary
          image/jpeg:
            schema:
              type: string
              description: payload
              format: binary
          image/webp:
            schema:
              type: string
              description: payload
              format: binary
          image/gif:
            schema:
              type: string
              description: payload
              format: binary
        required: true
      responses:
        '201':
          description: Info about uploaded image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfilePictureOk'
        '400':
          description: Payload is unsupported/corrupted or the hash (if provided) does not match the payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '411':
          description: Length is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '413':
          description: Image dimensions or byte size is too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
  /api/v1/account/email/verify:
    post:
      tags:
        - Account
      summary: Verify the users email address
      operationId: Verify Email Address
      parameters:
        - name: token
          in: query
          description: ''
          schema:
            type: string
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
components:
  schemas:
    AccountDto:
      required:
        - acceptedPrivacyPolicyVersion
        - acceptedTermsOfServiceVersion
        - createdAt
        - emailVerified
        - friends
        - id
        - lastOnline
        - obscuredEmail
        - profilePictureUrl
        - ssoConnections
        - status
        - statusText
        - updatedAt
        - username
      type: object
      properties:
        id:
          type: string
          description: ''
          format: uuid
        username:
          type: string
          description: ''
          example: MyUsername
        obscuredEmail:
          type: string
          description: Obfuscated email of your account
        emailVerified:
          type: boolean
          description: ''
        acceptedPrivacyPolicyVersion:
          type: integer
          description: ''
          format: int32
        acceptedTermsOfServiceVersion:
          type: integer
          description: ''
          format: int32
        profilePictureUrl:
          type: string
          description: ''
          format: uri
        status:
          allOf:
            - $ref: '#/components/schemas/UserStatus'
          description: ''
        statusText:
          type: string
          description: ''
        friends:
          type: array
          items:
            type: string
            format: uuid
          description: Id of friends this account has
        ssoConnections:
          type: array
          items:
            type: string
          description: SSO providers this account is connected to
        createdAt:
          type: string
          description: Date this user was created at
          format: date-time
        updatedAt:
          type: string
          description: Last time this user was updated
          format: date-time
        lastOnline:
          type: string
          description: Last time this user was online
          format: date-time
      additionalProperties: false
      description: 'Account object, this can only retrieved for the user you are logged in as'
...
    CreateAccount:
      required:
        - acceptedPrivacyPolicyVersion
        - acceptedTermsOfServiceVersion
        - email
        - password
        - turnstileResponse
        - username
      type: object
      properties:
        username:
          maxLength: 32
          minLength: 3
          pattern: '^[^\s].*[^\s]$'
          type: string
          description: ''
          example: MyUsername
        password:
          maxLength: 256
          minLength: 10
          type: string
          description: ''
          example: Hq2yP1B^Fho&zRHxHkEu
        email:
          type: string
          description: ''
        acceptedPrivacyPolicyVersion:
          type: integer
          description: ''
          format: int32
        acceptedTermsOfServiceVersion:
          type: integer
          description: ''
          format: int32
        turnstileResponse:
          type: string
          description: Response from cloudflare turnstile request
        ssoToken:
          type: string
          description: 'SSO token from OAuth flow, this is optional and only used when creating an account from the OAuth flow'
          nullable: true
      additionalProperties: false
      description: Message sent to server to create a new account
    CreateOk:
      required:
        - accountId
        - emailVerificationRequired
      type: object
      properties:
        accountId:
          type: string
          description: The account id of the newly created account
          format: uuid
        session:
          allOf:
            - $ref: '#/components/schemas/SessionDto'
          description: 'If the account was created using OAuth and in a way that ensures the email address is valid, a session will be created and returned here'
          nullable: true
        emailVerificationRequired:
          type: boolean
          description: If true then the email is already verified by 3rd party
      additionalProperties: false
    ErrorDetails:
      required:
        - code
        - detail
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        suggestion:
          type: string
          nullable: true
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
            nullable: true
          nullable: true
        notification:
          allOf:
            - $ref: '#/components/schemas/UserNotification'
          nullable: true
      additionalProperties: false
...
    NotificationSeverityLevel:
      enum:
        - info
        - warning
        - error
      type: string
...
    RecoveryConfirm:
      required:
        - newPassword
        - token
      type: object
      properties:
        newPassword:
          maxLength: 256
          minLength: 10
          type: string
          description: The new password to set
          example: Hq2yP1B^Fho&zRHxHkEu
        token:
          type: string
          description: The token sent to the user's email address
      additionalProperties: false
      description: Request sent to server to commit a password reset
...
    SessionDto:
      required:
        - expiresAtUtc
        - issuedAtUtc
        - sessionToken
      type: object
      properties:
        sessionToken:
          type: string
          format: uuid
        issuedAtUtc:
          type: string
          format: date-time
        expiresAtUtc:
          type: string
          format: date-time
      additionalProperties: false
...
    UpdateEmail:
      required:
        - newEmail
        - password
      type: object
      properties:
        newEmail:
          type: string
          description: ''
        password:
          type: string
          description: ''
          example: Hq2yP1B^Fho&zRHxHkEu
      additionalProperties: false
      description: Request sent to server to update account email address
    UpdateEmailOk:
      required:
        - message
      type: object
      properties:
        message:
          type: string
          description: 'Example: "Please check your email to verify your new address."'
      additionalProperties: false
      description: Request sent to server to update account email address
...
    UpdateProfilePictureOk:
      required:
        - imageHash
        - imageId
        - imageUrl
      type: object
      properties:
        imageId:
          type: string
          description: ''
          format: uuid
        imageUrl:
          type: string
          description: ''
          format: uri
        imageHash:
          type: string
          description: ''
      additionalProperties: false
...
    UserNotification:
      required:
        - content
        - severity
      type: object
      properties:
        severity:
          allOf:
            - $ref: '#/components/schemas/NotificationSeverityLevel'
        content:
          type: string
      additionalProperties: false
...
    UserStatus:
      enum:
        - offline
        - doNotDisturb
        - inactive
        - online
        - downBad
      type: string
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme.\r\n\r\n\r\n                      Enter 'Bearer' [space] and then your token in the text input below.\r\n                      \r\n\r\n                        Example: 'Bearer 12345abcdef'"
      name: Authorization
      in: header
security:
  - Bearer: [ ]